Implementando classe Album

+++ ScreenSound/Musica.cs
@@ -2,7 +2,7 @@
 { 
     public string Nome { get; set; }
     public string Artista { get; set; }
-    public double Duracao { get; set; }
+    public int Duracao { get; set; }
     public bool Disponivel { get; set; }
     public string DescricaoResumida => $"A música {Nome} pertence à banda {Artista}.";
